app/requirements.txt
amqp==2.5.2
argh==0.26.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.3.0
Django==3.0
djangorestframework==3.10.3
gunicorn==20.0.4
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
pathtools==0.1.2
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.2
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
watchdog==0.9.0
zipp==0.6.0
